LayoutTests/js/basic-for-of-expected.txt

This test checks the behavior of the for-of construct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS value is testArray[0]
PASS value is testArray[1]
PASS value is testArray[2]
PASS value is testArray[3]
PASS value is testArray[4]
PASS value is testArray[5]
PASS testArray.length is 6
PASS key is 0
PASS key is 1
PASS key is 2
PASS key is 3
PASS key is 4
PASS key is 5
PASS testArray.length is 6
PASS this.prop1 is 0
PASS this.prop1 is 1
PASS this.prop1 is 2
PASS this.prop1 is 3
PASS this.prop1 is 4
PASS this.prop1 is 5
PASS testArray.length is 6
PASS this[prop2] is 0
PASS this[prop2] is 1
PASS this[prop2] is 2
PASS this[prop2] is 3
PASS this[prop2] is 4
PASS this[prop2] is 5
PASS testArray.length is 6
PASS key is 0
PASS value is 1
PASS key is 1
PASS value is 2
PASS key is 2
PASS value is 3
PASS key is 3
PASS value is 4
PASS key is 4
PASS value is 5
PASS key is 5
PASS value is 6
PASS testArray.length is 6
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS testArray.length is 9
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS value is testArray[key]
PASS key is i
PASS testArray.length is 13
PASS successfullyParsed is true

TEST COMPLETE
